Column ids hopefully FINALLY FIXED, added start menu
added logarithmic timing for drops

fixed column_ids again.. hopefully for the last time. sometimes the
array would be accessed outside of its size due to some technicalities.
this would lead to corruption.

added a main menu, added all the mainmenu class with it's fonts and
displays. as well as instructions

oh yeah, added a class to hold all the resources for the puzzle game as
well called TimedResourceHolder.